+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Added
 - Single-Site DMRG with the :class:`~tenpy.algorithms.dmrg.SingleSiteDMRG`.
 - Example function in ``examples/c_tebd.py`` how to run TEBD with a model originally having next-nearest neighbors.
 - :meth:`~tenpy.networks.mps.MPS.increase_L` to allow increasing the unit cell of an MPS.
+- Additional option ``order='folded'`` for the :class:`~tenpy.models.lattice.Chain`.
 
 Fixed
 ^^^^^